Migrate applyEnforceNavigationBarContrast into expo-navigation-bar plugin
